server/requirements.txt
awscli==1.16.162
beautifulsoup4==4.7.1
botocore==1.12.152
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
cssselect==1.0.3
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.3.3
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.1
Pillow==6.0.0
pyasn1==0.4.5
python-dateutil==2.8.0
PyYAML==3.13
requests==2.22.0
requests-file==1.4.3
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
soupsieve==1.9.1
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.3
Werkzeug==0.15.4
